+++ NomaiVR/NomaiVR.csproj
@@ -138,6 +138,7 @@
     <Compile Include="ReusableBehaviours\Holdable.cs" />
     <Compile Include="ReusableBehaviours\PersistObject.cs" />
     <Compile Include="ReusableBehaviours\SmoothFollowCameraRotation.cs" />
+    <Compile Include="Tools\HolsterToolSwapper.cs" />
     <Compile Include="UI\ControllerModels.cs" />
     <Compile Include="UI\Dialogue.cs" />
     <Compile Include="Player\FeetMarker.cs" />
